move the api methods for fetching a session into the concern
