+++ lib/common/client/concerns/mhv_session_based_client.rb
@@ -18,6 +18,15 @@ module Common
         self
       end
 
+      def get_session
+        env = perform(:get, 'session', nil, auth_headers)
+        req_headers = env.request_headers
+        res_headers = env.response_headers
+        @session.class.new(user_id: req_headers['mhvCorrelationId'],
+                              expires_at: res_headers['expires'],
+                              token: res_headers['token'])
+      end
+
       module ClassMethods
         def client_session(klass = nil)
           @client_session ||= klass
