+++ lib/sm/client.rb
@@ -4,7 +4,6 @@ require 'common/client/base'
 require 'common/client/concerns/mhv_session_based_client'
 require 'sm/client_session'
 require 'sm/configuration'
-require 'sm/api/sessions'
 require 'sm/api/triage_teams'
 require 'sm/api/folders'
 require 'sm/api/messages'
@@ -13,7 +12,6 @@ require 'sm/api/message_drafts'
 module SM
   class Client < Common::Client::Base
     include Common::Client::MHVSessionBasedClient
-    include SM::API::Sessions
     include SM::API::TriageTeams
     include SM::API::Folders
     include SM::API::Messages
